backend/requirements.txt
fastapi
uvicorn
python-multipart
assemblyai
python-dotenv
pydantic
